third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/manifest-src-blocked-expected.txt

CONSOLE ERROR: Loading a manifest from 'http://127.0.0.1:8000/security/contentSecurityPolicy/manifest.test/manifest.json' violates the following Content Security Policy directive: "manifest-src 'none'". The action has been blocked.
ALERT: Pass
